Updated game of life
Updated game of life, today I tested functions that use PIL to both
write JPGs based on an array like the one in Game of Life, and read
JPGs to create a saved board. Eventually, both these will be
incorporated into the main game of life with a GUI, so that patterns
can be easily saved and recalled for the long term.